pyproject.toml
[project]
name = "hass-tryfi"
version = "0.0.25"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.2"
dependencies = [
    "homeassistant==2024.10.0",
]

[tool.ruff.lint]
extend-select = [
    "T201", # Print statements
    "D104" # undocumented __init__
]

[tool.ruff.lint.per-file-ignores]
"**/{tests,docs,tools}/*" = ["D104"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "requests-mock>=1.12.1",
    "responses>=0.25.8",
    "ruff>=0.12.2",
    "sentry-sdk>=2.32.0",
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "pytest-homeassistant-custom-component",
    "pytest-timeout",
]
